requirements.txt
django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0

channels==2.4.0
channels_redis==2.4.2
daphne==2.4.1

redis==3.4.1
coverage==5.0.3

requests==2.23.0
python-dotenv==0.12.0